MDL-56006 theme: Set Boost as the default theme
Units tests have been adjusted e.g. 'boost' instead of 'clean' in pix urls.

Some places I didn't update the test to reflect the boost markup and reset
the theme to clean. This is because the test was not testing the markup - but the
functionality (e.g. looking for specific structures to indicate an error in a form element).